+++ src/Heddle.Generator/HeddleTemplateGenerator.cs
@@ -484,7 +484,7 @@ namespace Heddle.Generator
                 }
             }
 
-            return "1.0.0";
+            return "2.0.0";
         }
     }
 }